Delete the entity framework
