--- frontend/src/metabase/entities/containers/rtk-query/index.ts
@@ -1,3 +1,0 @@
-export * from "./EntityListLoader";
-export * from "./EntityObjectLoader";
-export * from "./types";
