--- frontend/src/metabase/entities/containers/rtk-query/types/index.ts
@@ -1,1 +1,0 @@
-export * from "./entities";
