--- frontend/src/metabase/entities/index.ts
@@ -1,74 +1,0 @@
-/*
- * # Entities abstract the interface between the back-end and the front-end.
- *
- * ## Endpoint requirements for entities:
- *
- * When fetching a list, each item of the list must include an `id` key/value pair.
- *
- * JSON must wrap response inside a `{ "data" : { …your data } }` structure.
- *
- * ## Required Properties:
- *
- * name:
- *   a string in plural form
- *   examples:
- *     "questions", "dashboards"
- *
- * path:
- *   a uri
- *     starting with "/api/"
- *     conventionally followed by the entity name in singular form
- *   examples:
- *     "/api/card", "/api/dashboard"
- *
- * ## Optional properties:
- *
- * api:
- *
- * here you can override the basic entity methods like `list`, `create`, `get`, `update`, `delete` (OR see `path` below)
- *
- * schema:
- *   normalizr schema
- *   default:
- *     `new schema.Entity(entity.name)`
- *
- * ## How to create a bare-bones entity
- *
- * Say we want to create a "books" entity, to be able to fetch a list of "books".
- *
- * Add the following line to `frontend/src/metabase/entities.index.js`:
- *
- *   export { default as books } from "./books"
- *
- * Create file `frontend/src/metabase/entities/books.js`
- *
- * Add the following to it:
- *
- *   import { createEntity } from "./utils";
-
- *   const Books = createEntity({
- *     name: "books",
- *     nameOne: "book",
- *     path: "/api/book",
- *   });
- *
- *   export default Books;
- *
- * ## How to consume an entity:
- *
- * Near the top of a container file, import the entity:
- *
- *   import Book from "metabase/entities/books";
- *
- * Near the bottom of the container file, add the entity to a `compose` statement:
- *
- *   export default _.compose(
- *     Book.loadList(),
- *     connect(mapStateToProps),
- *   )(BookContainer);
- */
-
-// All entities have been migrated to "metabase/api"; this barrel intentionally
-// exports nothing. Keep the empty export so the file stays a module, not a
-// script.
-export {};
